Move `ImageCommentsUIComposer to production`
